+++ src/util/warn.ts
@@ -1,0 +1,3 @@
+export default function warn(message: string): void {
+  return console && console.warn(message);
+}
